Rearrange Approximator struct and related files
