+++ src/ReinforcementLearningCore/test/policies/learners/learners.jl
@@ -1,4 +1,2 @@
-include("abstract_learner.jl")
 include("approximator.jl")
 include("tabular_approximator.jl")
-include("target_network.jl")
