+++ src/ReinforcementLearningCore/test/policies/policies.jl
@@ -1,3 +1,4 @@
 include("agent.jl")
 include("multi_agent.jl")
 include("approximators.jl")
+include("learners.jl")
